:truck: Move modal.py under controlnet_ui
